# BROOKLYN-382: get-catalog performance improvements