--- a/server-cli/src/main/java/org/apache/brooklyn/cli/lister/ItemDescriptors.java
+++ b/server-cli/src/main/java/org/apache/brooklyn/cli/lister/ItemDescriptors.java
@@ -181,10 +181,9 @@ public class ItemDescriptors {
         return resolver.getPrefix();
     }
     
-    @SuppressWarnings({ "rawtypes", "unchecked" }) 
-    public static Map<String, Object> toItemDescriptor(BrooklynCatalog catalog, CatalogItem item, boolean headingsOnly) {
+    public static Map<String, Object> toItemDescriptor(BrooklynCatalog catalog, CatalogItem<?, ?> item, boolean headingsOnly) {
         Map<String,Object> itemDescriptor = MutableMap.of();
-        AbstractBrooklynObjectSpec<?,?> spec = catalog.createSpec(item);
+        AbstractBrooklynObjectSpec<?,?> spec = catalog.peekSpec(item);
         List<EntityConfigSummary> config = new ArrayList<>();
 
         if (item.getDisplayName() != null) {