… camp/camp-brooklyn/src/test/java/org/apache/brooklyn/camp/brooklyn/catalog/SpecParameterUnwrappingTest.java
… core/src/test/java/org/apache/brooklyn/core/mgmt/internal/EntityExecutionManagerTest.java
… core/src/test/java/org/apache/brooklyn/core/test/qa/performance/AbstractPerformanceTest.java
… core/src/test/java/org/apache/brooklyn/core/test/qa/performance/EntityPerformanceLongevityTest.java
… core/src/test/java/org/apache/brooklyn/core/test/qa/performance/EntityPerformanceTest.java
… core/src/test/java/org/apache/brooklyn/core/test/qa/performance/GroovyYardStickPerformanceTest.groovy
… core/src/test/java/org/apache/brooklyn/core/test/qa/performance/TaskPerformanceTest.java
… core/src/test/java/org/apache/brooklyn/util/core/internal/ssh/cli/SshCliToolPerformanceTest.java
… rest/rest-resources/src/test/java/org/apache/brooklyn/rest/resources/BrooklynRestResourcePerformanceTest.java
… rest/rest-resources/src/test/java/org/apache/brooklyn/rest/resources/CatalogResourcePerformanceTest.java
… rest/rest-resources/src/test/java/org/apache/brooklyn/rest/testing/BrooklynRestResourceTest.java
… test-support/src/main/java/org/apache/brooklyn/test/performance/PerformanceMeasurer.java
… test-support/src/main/java/org/apache/brooklyn/test/performance/PerformanceTestDescriptor.java