The `specCache` speeds up the REST api calls to `/catalog/entities` considerably.

For example, with 10 small yaml entities in the catalog, before the cache it takes about 170ms per call (and proportionately more as we increase the number of catalog items or their size):

```
test=PerformanceTestDescriptor{summary=CatalogResourcePerformanceTest.testListAllEntities, iterations=100, warmupIterations=10, job=org.apache.brooklyn.rest.resources.CatalogResourcePerformanceTest$8@246f150f, minAcceptablePerSecond=1.0}; result=PerformanceTestResult{duration=16s 981ms, warmup=3s 274ms, iterations=100, warmupIterations=10, ratePerSecond=5.888934691714269, histogram={134.2ms-268.4ms: 100}, cpuTotalFraction=1.3575868461219567, cpuSampleFractions=[1.2573016983016982, 1.9225778443113772, 1.4971834496510468, 1.5281373134328358, 1.1606623505976095, 1.2851962151394423, 1.2523649052841475, 1.2795744255744255, 1.211623880597015, 1.1832784431137724, 1.162554, 1.3302412761714855, 1.1046873126873127, 1.4044291417165669, 1.0696699900299103, 1.846793, 1.0917181274900398], minAcceptablePerSecond=1.0}
```

With the cache, it takes 24ms per call:
```
test=PerformanceTestDescriptor{summary=CatalogResourcePerformanceTest.testListAllEntities, iterations=100, warmupIterations=10, job=org.apache.brooklyn.rest.resources.CatalogResourcePerformanceTest$8@345a656a, minAcceptablePerSecond=1.0}; result=PerformanceTestResult{duration=2s 417ms, warmup=1s 393ms, iterations=100, warmupIterations=10, ratePerSecond=41.373603640877114, histogram={8.389ms-16.78ms: 20, 16.78ms-33.55ms: 74, 33.55ms-67.11ms: 4, 67.11ms-134.2ms: 1, 134.2ms-268.4ms: 1}, cpuTotalFraction=2.552031110340871, cpuSampleFractions=[2.367437, 1.722136], minAcceptablePerSecond=1.0}
```