@neykov 2016-11-10
Need to clear the **whole** cache on any addition/removal/update. Specs will change when transient catalog item dependencies change.
There are things outside of the control of the catalog that a spec depends on - like non-catalog locations, type registry, adding bundles.
Re-using the specs will cause objects to be shared between applications - something that could lead to subtle problems (similar to the dsl memberSpeck locking).

Having said that I believe `BrooklynCatalog.createSpec` is only used by the REST API so not as bad as it sounds. But better move the logic to the REST API?

The cache will speed up listing, but not adding items or deploying which could still take considerable time. A much better fix would be to streamline our catalog parsing logic which currently is way too complicated.