@aledsage 2016-11-14
@neykov great comments - I've addressed them. Can you take another look please?