@aledsage 2016-11-15
The commit `EntityExecutionManagerTest: tidy` is because `org.apache.brooklyn.core.mgmt.internal.EntityExecutionManagerTest.testGcTaskAtNormalTagLimit` failed in the jenkins PR build. But I can't reproduce that, and I don't have the full stacktrace unfortunately.