@neykov 2016-11-16
Looks good, some comments to consider.