@aledsage 2016-11-16
@neykov I believe those comments are now addressed (note the "Performance" profile stuff has been removed from this PR, pending discussion on dev@brooklyn).

Are you happy with merging this now?