@neykov 2016-11-16
Thanks @aledsage. Looks good, merging.